+++ src/main/java/com/notenoughmail/kubejs_tfc/util/helpers/SizePredicate.java
@@ -1,0 +1,9 @@
+package com.notenoughmail.kubejs_tfc.util.helpers;
+
+import net.dries007.tfc.common.capabilities.size.Size;
+
+@FunctionalInterface
+public interface SizePredicate {
+
+    boolean test(Size size);
+}