+++ src/main/java/com/notenoughmail/kubejs_tfc/util/helpers/WeightPredicate.java
@@ -1,0 +1,9 @@
+package com.notenoughmail.kubejs_tfc.util.helpers;
+
+import net.dries007.tfc.common.capabilities.size.Weight;
+
+@FunctionalInterface
+public interface WeightPredicate {
+
+    boolean test(Weight weight);
+}