+++ src/main/java/com/notenoughmail/kubejs_tfc/util/helpers/ducks/IKubeJSTFCHandHeldItemBuilderExtension.java
@@ -1,4 +1,4 @@
-package com.notenoughmail.kubejs_tfc.util.implementation.ducks;
+package com.notenoughmail.kubejs_tfc.util.helpers.ducks;
 
 import dev.latvian.mods.kubejs.item.MutableToolTier;
 import dev.latvian.mods.kubejs.item.custom.HandheldItemBuilder;