+++ src/main/java/com/notenoughmail/kubejs_tfc/util/implementation/event/RegisterRepresentativeBlocksEventJS.java
@@ -5,6 +5,7 @@ import dev.latvian.mods.kubejs.registry.RegistryInfo;
 import dev.latvian.mods.kubejs.typings.Generics;
 import dev.latvian.mods.kubejs.typings.Info;
 import dev.latvian.mods.kubejs.typings.Param;
+import dev.latvian.mods.kubejs.util.UtilsJS;
 import net.dries007.tfc.common.items.PropickItem;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.Block;
@@ -27,10 +28,10 @@ public class RegisterRepresentativeBlocksEventJS extends EventJS {
     @Generics(value = ResourceLocation.class)
     public void registerRepresentative(ResourceLocation representative, List<ResourceLocation> blocks) {
         final Block rep = RegistryInfo.BLOCK.getValue(representative);
-        final List<Block> list = new ArrayList<>();
-        for (ResourceLocation block : blocks) {
-            list.add(RegistryInfo.BLOCK.getValue(block));
+        final Block[] array = new Block[blocks.size()];
+        for (int i = 0 ; i < blocks.size() ; i++) {
+            array[i] = RegistryInfo.BLOCK.getValue(UtilsJS.getMCID(null, blocks.get(i)));
         }
-        PropickItem.registerRepresentative(rep, list.toArray(new Block[0]));
+        PropickItem.registerRepresentative(rep, array);
     }
 }